backend/requirements.txt
python-dotenv
# Django Core (Django 4.2 LTS - supports Python 3.9+)
Django>=4.2,<5.0
python-dotenv
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0

# AWS S3
boto3>=1.28.0
django-storages>=1.14.0

# API Documentation
drf-spectacular>=0.27.0

# Nested routers
drf-nested-routers>=0.93.0

# Environment
python-decouple>=3.8

# Utilities
Pillow>=9.0.0
cryptography==46.0.3
celery==5.6.2
redis==7.1.0
# Development (optional)
ipython>=8.0.0

# Testing (optional)
pytest>=7.0.0
pytest-django>=4.5.0

# Code Quality (optional)
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
